+++ src/label/components.js
@@ -2,31 +2,7 @@ import { $, request } from '../utils';
 import { getLabelItemTpl, getLabelTpl } from '../tpl.js';
 import { labelFormStore, labelListStore } from '../store/labelStore.js';
 
-export class LabelButton {
-  addEvent() {
-    const labelBtn = $('#label-btn');
-    labelBtn.addEventListener('click', () => {
-      $('#app').innerHTML = getLabelTpl();
-      this.initLabelPage();
-    });
-  }
-  initLabelPage() {
-    const labelForm = LabelForm();
-    labelForm.init();
-    new LabelList({
-      setList: labelListStore.setLabelList,
-      subscribe: labelListStore.subscribe,
-    });
-    new NewLabelBtn({
-      subscribe: labelFormStore.subscribe,
-      toggleFormOpened: labelFormStore.toggleFormOpened,
-      revealForm: labelForm.revealForm,
-      hideForm: labelForm.hideForm,
-    });
-  }
-}
-
-const LabelForm = (selector = '#new-label-form') => {
+export const LabelForm = (selector = '#new-label-form') => {
   const $target = $(selector);
   return {
     $target,
