+++ src/main.js
@@ -1,9 +1,9 @@
 import setIssueOnDocument from './issue';
-import { Label } from './label/Label.js';
+import { LabelButton } from './label/Label.js';
 
 import { worker } from './mocks/browser.js';
 worker.start();
 
 setIssueOnDocument();
 
-new Label();
+new LabelButton().addEvent();
